files and variables naming fix
