--- src/components/SubTask/SubTaskProgressBar/subTaskProgressBarUtils.js
@@ -1,19 +1,0 @@
-export const getProgressBarVariant = (percent = 0) => {
-  switch (true) {
-    case percent === 100:
-      return 'success';
-    case percent > 50:
-      return 'info';
-    case percent > 30:
-      return 'warning';
-    default:
-      return 'secondary';
-  }
-};
-
-export const getCompletedSubTasksPercent = (subtaskList = {}) => {
-  if (!Object.keys(subtaskList).length) return 0;
-  const completedSubTasks = subtaskList.filter(subTask => subTask.completed);
-  const completedSubTasksPercent = (completedSubTasks.length / subtaskList.length) * 100;
-  return Math.round(completedSubTasksPercent);
-};
