+++ src/components/TaskDetails/TaskDetailsModal.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Modal, Button } from 'react-bootstrap';
 import TaskDetails from './TaskDetails';
-import SubTaskListContainer from '../SubTask/SubTaskListContainer';
+import SubtaskListContainer from '../Subtask/SubtaskListContainer';
 import DeleteTask from './DeleteTask';
 
 export default class TaskDetailsModal extends React.Component {
@@ -39,7 +39,7 @@ export default class TaskDetailsModal extends React.Component {
             taskRef={taskRef}
             closeTaskDetails={this.closeTaskDetails}
           />
-          <SubTaskListContainer taskRef={taskRef} />
+          <SubtaskListContainer taskRef={taskRef} />
         </Modal.Body>
         <Button
           variant="outline-danger"
