--- src/components/__tests__/SubTaskFilterButton.test.js
@@ -1,62 +1,0 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-import { SubTaskFilterButtonComponent } from '../SubTask/SubTaskFilter/SubTaskFilterButton';
-import { SHOW_ALL, SHOW_ACTIVE, SHOW_COMPLETED } from '../SubTask/subTaskFilterTypes';
-
-describe('SubTaskFilterButton component', () => {
-  const subTaskActions = { setSubTaskFilter: jest.fn() };
-  it('should render correctly and match the snapshot', () => {
-    const subTaskFilterButtonComponent = shallow(
-      <SubTaskFilterButtonComponent subTaskActions={subTaskActions} />,
-    );
-    expect(subTaskFilterButtonComponent).toMatchSnapshot();
-  });
-
-  it('should render Bootstrap 4 Button', () => {
-    const subTaskFilterButtonComponent = shallow(
-      <SubTaskFilterButtonComponent subTaskActions={subTaskActions} />,
-    );
-    expect(subTaskFilterButtonComponent.find('Button').length).toBe(1);
-  });
-
-  it('should render Bootstrap 4 Button with text depending on "filterType" prop', () => {
-    const subTaskFilterButtonComponent = shallow(
-      <SubTaskFilterButtonComponent
-        subTaskActions={subTaskActions}
-        filterType={SHOW_ALL}
-      />,
-    );
-    expect(subTaskFilterButtonComponent.find('Button').text()).toBe('All');
-  });
-
-  it('should render Bootstrap 4 Button with text depending on "filterType" prop', () => {
-    const subTaskFilterButtonComponent = shallow(
-      <SubTaskFilterButtonComponent
-        subTaskActions={subTaskActions}
-        filterType={SHOW_ACTIVE}
-      />,
-    );
-    expect(subTaskFilterButtonComponent.find('Button').text()).toBe('Active');
-  });
-
-  it('should render Bootstrap 4 Button with text depending on "filterType" prop', () => {
-    const subTaskFilterButtonComponent = shallow(
-      <SubTaskFilterButtonComponent
-        subTaskActions={subTaskActions}
-        filterType={SHOW_COMPLETED}
-      />,
-    );
-    expect(subTaskFilterButtonComponent.find('Button').text()).toBe('Completed');
-  });
-
-  it('should render active Bootstrap 4 Button if prop "filterType" is equal to prop "currentFilter"', () => {
-    const subTaskFilterButtonComponent = shallow(
-      <SubTaskFilterButtonComponent
-        subTaskActions={subTaskActions}
-        filterType={SHOW_COMPLETED}
-        currentFilter={SHOW_COMPLETED}
-      />,
-    );
-    expect(subTaskFilterButtonComponent.find('Button').prop('active')).toBeTruthy();
-  });
-});
